+++ src/main/java/org/influxdb/impl/InfluxDBMapper.java
@@ -21,6 +21,11 @@ public class InfluxDBMapper extends InfluxDBResultMapper {
     this.influxDB = influxDB;
   }
 
+  public <T> List<T> query(final Query query, final Class<T> clazz, final String measurementName) {
+    QueryResult queryResult = influxDB.query(query);
+    return toPOJO(queryResult, clazz, measurementName);
+  }
+
   public <T> List<T> query(final Query query, final Class<T> clazz) {
     throwExceptionIfMissingAnnotation(clazz);
     QueryResult queryResult = influxDB.query(query);